+++ src/features/keytar/user-token.ts
@@ -1,4 +1,3 @@
-import keytar from 'keytar';
 import os from 'node:os';
 import { setResponseKO, setResponseOK, slugify } from 'oro-functions';
 import type { SResponseKOSimple, SResponseOKBasic } from 'oro-functions';
@@ -7,11 +6,14 @@ import { OGH_PROJECT } from '@/features/global/constants.js';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import type { UnknownAny } from '@/features/global/types.js';
 
+import { getKeytar } from './get-keytar.js';
+
 //
 
 export async function getUserToken() {
   try {
     const username = slugify(os.userInfo().username ?? 'unknown');
+    const keytar = await getKeytar();
 
     const token = await keytar.getPassword(OGH_PROJECT, username);
 
@@ -26,6 +28,7 @@ export async function getUserToken() {
 export async function saveUserToken(token: string): Promise<SResponseOKBasic | SResponseKOSimple> {
   try {
     const username = slugify(os.userInfo().username ?? 'unknown');
+    const keytar = await getKeytar();
 
     await keytar.setPassword(OGH_PROJECT, username, token);
 
@@ -40,6 +43,7 @@ export async function saveUserToken(token: string): Promise<SResponseOKBasic | S
 export async function deleteUserToken() {
   try {
     const username = slugify(os.userInfo().username ?? 'unknown');
+    const keytar = await getKeytar();
 
     await keytar.deletePassword(OGH_PROJECT, username);